refactor: qc-application user 도메인 패키지 정리
