+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/account/controller/UserAccountController.java
@@ -1,7 +1,9 @@
-package com.eager.questioncloud.application.api.user.account;
+package com.eager.questioncloud.application.api.user.account.controller;
 
 import com.eager.questioncloud.application.api.common.DefaultResponse;
-import com.eager.questioncloud.application.api.user.account.UserAccountControllerResponse.RecoverForgottenEmailResponse;
+import com.eager.questioncloud.application.api.user.account.dto.UserAccountControllerRequest;
+import com.eager.questioncloud.application.api.user.account.dto.UserAccountControllerResponse.RecoverForgottenEmailResponse;
+import com.eager.questioncloud.application.api.user.account.service.UserAccountService;
 import com.eager.questioncloud.application.security.UserPrincipal;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
