+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/coupon/controller/UserCouponController.java
@@ -1,7 +1,9 @@
-package com.eager.questioncloud.application.api.user.coupon;
+package com.eager.questioncloud.application.api.user.coupon.controller;
 
 import com.eager.questioncloud.application.api.common.DefaultResponse;
-import com.eager.questioncloud.application.api.user.coupon.UserCouponControllerResponse.GetAvailableUserCouponsResponse;
+import com.eager.questioncloud.application.api.user.coupon.dto.UserCouponControllerRequest;
+import com.eager.questioncloud.application.api.user.coupon.dto.UserCouponControllerResponse.GetAvailableUserCouponsResponse;
+import com.eager.questioncloud.application.api.user.coupon.service.UserCouponService;
 import com.eager.questioncloud.application.security.UserPrincipal;
 import com.eager.questioncloud.core.domain.coupon.dto.AvailableUserCoupon;
 import io.swagger.v3.oas.annotations.Operation;
