+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/coupon/service/UserCouponService.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.application.api.user.coupon;
+package com.eager.questioncloud.application.api.user.coupon.service;
 
+import com.eager.questioncloud.application.api.user.coupon.implement.UserCouponRegister;
 import com.eager.questioncloud.core.domain.coupon.dto.AvailableUserCoupon;
 import com.eager.questioncloud.core.domain.coupon.infrastructure.UserCouponRepository;
 import com.eager.questioncloud.lock.LockKeyGenerator;
