+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/point/controller/UserPointController.java
@@ -1,6 +1,7 @@
-package com.eager.questioncloud.application.api.user.point;
+package com.eager.questioncloud.application.api.user.point.controller;
 
-import com.eager.questioncloud.application.api.user.point.UserPointControllerResponse.GetUserPointResponse;
+import com.eager.questioncloud.application.api.user.point.dto.UserPointControllerResponse.GetUserPointResponse;
+import com.eager.questioncloud.application.api.user.point.service.UserPointService;
 import com.eager.questioncloud.application.security.UserPrincipal;
 import com.eager.questioncloud.core.domain.point.model.UserPoint;
 import io.swagger.v3.oas.annotations.Operation;
