+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/point/service/UserPointService.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.user.point;
+package com.eager.questioncloud.application.api.user.point.service;
 
 import com.eager.questioncloud.core.domain.point.infrastructure.UserPointRepository;
 import com.eager.questioncloud.core.domain.point.model.UserPoint;
