+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/profile/dto/UserPointControllerResponse.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.user.profile;
+package com.eager.questioncloud.application.api.user.profile.dto;
 
 import com.eager.questioncloud.core.domain.user.dto.MyInformation;
 import lombok.AllArgsConstructor;
